Move ManualWithSections#remove_section into Manual
I've moved the corresponding tests from `ManualWithSections` spec into
`Manual` spec almost unchanged. This has then allowed me to remove the
now empty `ManualWithSections` spec.
