+++ app/models/manual_with_sections.rb
@@ -6,18 +6,6 @@ class ManualWithSections
     @removed_sections = removed_sections
   end
 
-  def remove_section(section_id)
-    found_section = sections.find { |d| d.id == section_id }
-
-    return if found_section.nil?
-
-    removed = sections.delete(found_section)
-
-    return if removed.nil?
-
-    removed_sections << removed
-  end
-
   def add_section(section)
     sections << section
   end
